distance_method/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
fonttools==4.54.1
frozendict==2.4.6
html5lib==1.1
idna==3.10
kiwisolver==1.4.7
lxml==5.3.0
matplotlib==3.9.2
multitasking==0.0.11
numpy==2.1.2
packaging==24.1
pandas==2.2.3
peewee==3.17.7
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
yfinance==0.1.70
mysqlclient==2.2.4
